Uses context for main action
