--- script/cli-test.sh
@@ -1,11 +1,0 @@
-#!/bin/bash
-set -e
-rm -rf report/shell-tests
-mkdir -p report/shell-tests
-BALDRICK_BROTH_BUILD_FILE=script/fixture/.baldrick-broth/dev.yaml yarn cli --help
-echo "---"
-BALDRICK_BROTH_BUILD_FILE=script/fixture/.baldrick-broth/dev.yaml yarn cli test --help
-echo "---"
-BALDRICK_BROTH_BUILD_FILE=script/fixture/.baldrick-broth/dev.yaml yarn cli test generate --help
-echo "---"
-BALDRICK_BROTH_BUILD_FILE=script/fixture/.baldrick-broth/dev.yaml yarn cli test generate -c blue
